requirements.txt
torch
numpy
scipy
PyYAML
librosa
SoundFile
inflect
tensorboardX
Unidecode
pillow